dbus-proxy: - split main C file into smaller modules - add built-in BlueZ and NetworkManager agent support - enable multiple agent registrations - remove nm-mode param and NM_SECRET_AGENT_XML env var - update NM/BlueZ agent handling (not fully tested)
