+++ packages/cpp/dbus-proxy/dbus-proxy/gdbusprivate.h
@@ -1,22 +1,29 @@
 /*
- Copyright 2025 TII (SSRC) and the Ghaf contributors
+ Copyright 2026 TII (SSRC) and the Ghaf contributors
  SPDX-License-Identifier: Apache-2.0
  */
+#ifndef GDBUS_PRIVATE_H
+#define GDBUS_PRIVATE_H
 
-constexpr const char *DBUS_INTERFACE_OBJECT_MANAGER =
+constexpr const gchar *DBUS_INTERFACE_OBJECT_MANAGER =
     "org.freedesktop.DBus.ObjectManager";
-constexpr const char *DBUS_SIGNAL_INTERFACES_ADDED = "InterfacesAdded";
-constexpr const char *DBUS_SIGNAL_INTERFACES_REMOVED = "InterfacesRemoved";
-constexpr const char *DBUS_OBJECT_PATH = "/org/freedesktop";
-constexpr const char *DBUS_OBJECT_PATH_DBUS = "/org/freedesktop/DBus";
-constexpr const char *DBUS_INTERFACE_INTROSPECTABLE =
+constexpr const gchar *DBUS_SIGNAL_INTERFACES_ADDED = "InterfacesAdded";
+constexpr const gchar *DBUS_SIGNAL_INTERFACES_REMOVED = "InterfacesRemoved";
+constexpr const gchar *DBUS_OBJECT_PATH_DBUS = "/org/freedesktop/DBus";
+constexpr const gchar *DBUS_INTERFACE_INTROSPECTABLE =
     "org.freedesktop.DBus.Introspectable";
-constexpr const char *DBUS_INTERFACE_PROPERTIES =
+constexpr const gchar *DBUS_INTERFACE_PROPERTIES =
     "org.freedesktop.DBus.Properties";
-constexpr const char *DBUS_INTERFACE_PEER = "org.freedesktop.DBus.Peer";
-constexpr const char *DBUS_INTERFACE_SECRET_AGENT =
-    "org.freedesktop.NetworkManager.SecretAgent";
-constexpr const char *DBUS_INTERFACE_NETWORK_MANAGER =
+constexpr const gchar *DBUS_INTERFACE_PEER = "org.freedesktop.DBus.Peer";
+
+constexpr const gchar *DBUS_NETWORK_MANAGER_NAME =
     "org.freedesktop.NetworkManager";
-constexpr const char *DBUS_OBJECT_PATH_NETWORK_MANAGER =
-    "/org/freedesktop/NetworkManager";
+constexpr const gchar *DBUS_INTERFACE_SECRET_AGENT =
+    "org.freedesktop.NetworkManager.SecretAgent";
+constexpr const gchar *DBUS_NM_AGENT_PATH =
+    "/org/freedesktop/NetworkManager/SecretAgent";
+
+constexpr const gchar *DBUS_BT_AGENT_PATH = "/org/bluez/agent";
+constexpr const gchar *DBUS_INTERFACE_BT_AGENT = "org.bluez.Agent1";
+
+#endif // GDBUS_PRIVATE_H
