+++ packages/cpp/dbus-proxy/default.nix
@@ -1,4 +1,4 @@
-# Copyright 2025 TII (SSRC) and the Ghaf contributors
+# Copyright 2026 TII (SSRC) and the Ghaf contributors
 # SPDX-License-Identifier: Apache-2.0
 {
   stdenv,
@@ -8,7 +8,7 @@
 }:
 stdenv.mkDerivation {
   pname = "dbus-proxy";
-  version = "0.1.0";
+  version = "0.1.2";
 
   src = ./dbus-proxy;
